Split app.rs into app/ submodule directory
state.rs: GpuMesh, MicrotomeApp struct and constructor
helpers.rs: mesh upload, model matrix, job control methods
ui_impl.rs: eframe::App trait impl with the main ui() method
defaults.rs: default printer and job config factories
