+++ crates/microtome-app/src/app/mod.rs
@@ -1,0 +1,8 @@
+//! Main application state and UI layout for Microtome.
+
+mod defaults;
+mod helpers;
+mod state;
+mod ui_impl;
+
+pub use state::MicrotomeApp;
