+++ crates/microtome-app/src/app/ui_impl.rs
@@ -1,260 +1,18 @@
-//! Main application state and UI layout for Microtome.
+//! eframe::App trait implementation for MicrotomeApp.
 
-use std::path::PathBuf;
 use std::sync::Arc;
-use std::sync::atomic::AtomicBool;
-use std::sync::mpsc;
 
 use glam::{Mat4, Quat, Vec3};
-use microtome_core::{
-    GpuContext, MeshData, PrintJobConfig, PrintMesh, PrintVolume, PrinterConfig, PrinterScene,
-    Projector, SliceProgress, ZStage, run_slicing_job,
-};
+use microtome_core::{GpuContext, MeshData, PrintMesh};
 use transform_gizmo_egui::math::Transform;
 use transform_gizmo_egui::prelude::*;
-use wgpu::util::DeviceExt;
 
-use crate::camera::OrbitCamera;
 use crate::slice_preview::SlicePreview;
 use crate::ui::panels::{self, AppState};
 use crate::viewport::ViewportPaintCallback;
-use crate::viewport_renderer::{MeshBuffers, ViewportRenderer};
+use crate::viewport_renderer::ViewportRenderer;
 
-/// GPU-side buffers corresponding to a scene mesh.
-struct GpuMesh {
-    /// Vertex buffer on the GPU.
-    vertex_buffer: wgpu::Buffer,
-    /// Index buffer on the GPU.
-    index_buffer: wgpu::Buffer,
-    /// Number of indices.
-    index_count: u32,
-}
-
-/// Main application state holding the scene, configuration, and UI state.
-pub struct MicrotomeApp {
-    /// The 3D print scene with volume and meshes.
-    scene: PrinterScene,
-    /// Hardware printer configuration.
-    printer_config: PrinterConfig,
-    /// Print job settings.
-    job_config: PrintJobConfig,
-    /// Orbit camera for viewport navigation.
-    camera: OrbitCamera,
-    /// Current slice plane Z height in mm.
-    slice_z: f32,
-    /// Index of the currently selected mesh, if any.
-    selected_mesh: Option<usize>,
-    /// GPU buffers for each mesh in the scene.
-    gpu_meshes: Vec<GpuMesh>,
-    /// Whether the render state was successfully initialized.
-    has_render_state: bool,
-    /// 2D slice preview panel.
-    slice_preview: SlicePreview,
-    /// Overhang angle in degrees for visualization.
-    overhang_angle_degrees: f32,
-    /// Active slicing job progress (0.0 to 1.0).
-    slicing_progress: Option<f32>,
-    /// Receiver for slicing progress updates from the background job.
-    progress_rx: Option<mpsc::Receiver<SliceProgress>>,
-    /// Path to save the ZIP output when slicing completes.
-    export_path: Option<PathBuf>,
-    /// Cancellation flag for active slicing job.
-    cancel_flag: Option<Arc<AtomicBool>>,
-    /// 3D transform gizmo for interactive manipulation.
-    gizmo: Gizmo,
-    /// Current gizmo modes (translate, rotate, or scale group).
-    gizmo_modes: EnumSet<GizmoMode>,
-    /// Whether to show the slice overlay in the 3D viewport.
-    show_slice_overlay: bool,
-    /// Snapshot of the volume config from last frame (for change detection).
-    prev_volume: PrintVolume,
-    /// Snapshot of the projector config from last frame (for change detection).
-    prev_projector: Projector,
-}
-
-impl MicrotomeApp {
-    /// Creates a new application with default configurations.
-    ///
-    /// Initializes the [`ViewportRenderer`] and stores it in egui-wgpu's
-    /// callback resources when a wgpu render state is available.
-    pub fn new(cc: &eframe::CreationContext) -> Self {
-        let printer_config = default_printer_config();
-        let scene = PrinterScene::from_config(&printer_config);
-        let mut has_render_state = false;
-
-        if let Some(render_state) = &cc.wgpu_render_state {
-            let mut renderer =
-                ViewportRenderer::new(&render_state.device, render_state.target_format);
-            renderer.update_volume_lines(&render_state.device, &scene.volume);
-            render_state
-                .renderer
-                .write()
-                .callback_resources
-                .insert(renderer);
-            has_render_state = true;
-        }
-
-        // Use half the projector resolution for responsive preview
-        let preview_width = printer_config.projector.x_res_px / 2;
-        let preview_height = printer_config.projector.y_res_px / 2;
-
-        let prev_volume = printer_config.volume.clone();
-        let prev_projector = printer_config.projector.clone();
-
-        Self {
-            scene,
-            printer_config,
-            job_config: default_job_config(),
-            camera: OrbitCamera::new(),
-            slice_z: 0.0,
-            selected_mesh: None,
-            gpu_meshes: Vec::new(),
-            has_render_state,
-            slice_preview: SlicePreview::new(preview_width, preview_height),
-            overhang_angle_degrees: 0.0,
-            slicing_progress: None,
-            progress_rx: None,
-            export_path: None,
-            cancel_flag: None,
-            gizmo: Gizmo::default(),
-            gizmo_modes: GizmoMode::all(),
-            show_slice_overlay: true,
-            prev_volume,
-            prev_projector,
-        }
-    }
-
-    /// Uploads a loaded mesh's vertex and index data to the GPU.
-    fn upload_mesh(&mut self, render_state: &egui_wgpu::RenderState, mesh_data: &MeshData) {
-        let vertex_buffer =
-            render_state
-                .device
-                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                    label: Some("mesh_vertices"),
-                    contents: bytemuck::cast_slice(&mesh_data.vertices),
-                    usage: wgpu::BufferUsages::VERTEX,
-                });
-
-        let index_buffer =
-            render_state
-                .device
-                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                    label: Some("mesh_indices"),
-                    contents: bytemuck::cast_slice(&mesh_data.indices),
-                    usage: wgpu::BufferUsages::INDEX,
-                });
-
-        self.gpu_meshes.push(GpuMesh {
-            vertex_buffer,
-            index_buffer,
-            index_count: mesh_data.indices.len() as u32,
-        });
-    }
-
-    /// Builds the model matrix for a [`PrintMesh`] from its position, rotation, and scale.
-    fn model_matrix(mesh: &PrintMesh) -> Mat4 {
-        let translation = Mat4::from_translation(mesh.position);
-        let rotation = Mat4::from_quat(Quat::from_euler(
-            glam::EulerRot::XYZ,
-            mesh.rotation.x,
-            mesh.rotation.y,
-            mesh.rotation.z,
-        ));
-        let scale = Mat4::from_scale(mesh.scale);
-        translation * rotation * scale
-    }
-
-    /// Builds the list of [`MeshBuffers`] for the current frame.
-    fn collect_mesh_buffers(&self) -> Vec<MeshBuffers> {
-        self.gpu_meshes
-            .iter()
-            .zip(self.scene.meshes.iter())
-            .map(|(gpu, scene_mesh)| MeshBuffers {
-                vertex_buffer: gpu.vertex_buffer.clone(),
-                index_buffer: gpu.index_buffer.clone(),
-                index_count: gpu.index_count,
-                model_matrix: Self::model_matrix(scene_mesh),
-            })
-            .collect()
-    }
-
-    /// Polls the progress receiver for slicing job updates.
-    ///
-    /// Handles progress fraction updates, job completion (writing ZIP to disk),
-    /// and job failure/cancellation.
-    fn poll_slicing_progress(&mut self) {
-        let should_clear = if let Some(rx) = &self.progress_rx {
-            let mut should_clear = false;
-            // Drain all available messages
-            while let Ok(msg) = rx.try_recv() {
-                match msg {
-                    SliceProgress::Progress(p) => {
-                        self.slicing_progress = Some(p);
-                    }
-                    SliceProgress::Complete(zip_bytes) => {
-                        if let Some(path) = &self.export_path {
-                            if let Err(e) = std::fs::write(path, &zip_bytes) {
-                                log::error!("Failed to write ZIP: {e}");
-                            } else {
-                                log::info!("Exported ZIP to: {}", path.display());
-                            }
-                        }
-                        should_clear = true;
-                        break;
-                    }
-                    SliceProgress::Failed(msg) => {
-                        log::error!("Slicing job failed: {msg}");
-                        should_clear = true;
-                        break;
-                    }
-                }
-            }
-            should_clear
-        } else {
-            false
-        };
-
-        if should_clear {
-            self.slicing_progress = None;
-            self.progress_rx = None;
-            self.export_path = None;
-            self.cancel_flag = None;
-        }
-    }
-
-    /// Starts a slicing job on a background thread.
-    ///
-    /// Constructs a fresh [`PrinterScene`] from the current config and meshes,
-    /// then spawns the job in a background thread.
-    fn start_slicing_job(&mut self) {
-        let (tx, rx) = mpsc::channel();
-        let cancel = Arc::new(AtomicBool::new(false));
-
-        // Build a fresh scene for the background thread (PrinterScene is not Clone)
-        let mut scene = PrinterScene::from_config(&self.printer_config);
-        for mesh in &self.scene.meshes {
-            scene.add_mesh(mesh.clone());
-        }
-        scene.set_overhang_angle_degrees(self.overhang_angle_degrees as f64);
-
-        let printer_config = self.printer_config.clone();
-        let job_config = self.job_config.clone();
-        let cancel_clone = Arc::clone(&cancel);
-        let tx_clone = tx.clone();
-
-        std::thread::spawn(move || {
-            if let Err(e) =
-                run_slicing_job(&scene, &printer_config, &job_config, tx_clone, cancel_clone)
-            {
-                let _ = tx.send(SliceProgress::Failed(e.to_string()));
-            }
-        });
-
-        self.slicing_progress = Some(0.0);
-        self.progress_rx = Some(rx);
-        self.cancel_flag = Some(cancel);
-    }
-}
+use super::state::MicrotomeApp;
 
 impl eframe::App for MicrotomeApp {
     /// Main UI rendering called each frame.
@@ -649,45 +407,5 @@ impl eframe::App for MicrotomeApp {
                 );
             }
         });
-    }
-}
-
-/// Returns the default printer configuration.
-fn default_printer_config() -> PrinterConfig {
-    PrinterConfig {
-        name: "Default Printer".into(),
-        description: String::new(),
-        last_modified: 0,
-        volume: PrintVolume {
-            width_mm: 120.0,
-            depth_mm: 68.0,
-            height_mm: 150.0,
-        },
-        z_stage: ZStage {
-            lead_mm: 8.0,
-            steps_per_rev: 200,
-            microsteps: 16,
-        },
-        projector: Projector {
-            x_res_px: 2560,
-            y_res_px: 1440,
-        },
-    }
-}
-
-/// Returns the default print job configuration.
-fn default_job_config() -> PrintJobConfig {
-    PrintJobConfig {
-        name: "Default Job".into(),
-        description: String::new(),
-        step_distance_microns: 2.5,
-        steps_per_layer: 20,
-        settle_time_ms: 3000,
-        layer_exposure_time_ms: 8000,
-        blank_time_ms: 500,
-        retract_distance_mm: 6.0,
-        z_offset_mm: 0.3,
-        raft_thickness_mm: 0.2,
-        raft_outset_mm: 0.5,
     }
 }
